requirements.txt
cryptography>=41.0.0
pyperclip>=1.8.2
pystray>=0.19.4
Pillow>=10.0.0
pynput>=1.8.0